fix(chairops/pos-ingest): coin event counters Int → BigInt + dedup note
CEO 2026-06-01: coin commit blew up with
  "Value out of range for the type: value \"4293916823\" is out of range
   for type integer"

StarThing's coin meter is an unsigned 32-bit counter and at least one
chair already crossed signed-int4 max. Bumped both columns to BIGINT
(Postgres int8) which covers 9.2e18 — 9 orders of magnitude of headroom.

- schema.prisma: ChairopsPosCoinEvent.coinAdded / coinMeter → BigInt
- prisma/migrations/20260601_coin_event_bigint.sql: ALTER TABLE for prod
  (metadata-only change on Postgres ≥10 · no rewrite · safe live)
- lib/chairops/pos-ingest/event-diff.ts: wrap Math.round() output in
  BigInt() for Prisma createMany input
- Cash side already uses Decimal(12,2) so no change needed there
- rowHash format unchanged (toString of the integer) → dedup still works
  identically across the schema change

UI · CEO ask "ตรวจสอบข้อมูลซ้ำด้วย":
- The dedup engine already keys on
  sha256(device | eventAt | amount | meter), so re-uploading day 1-3
  then 3-4 will correctly skip day-3 rows (bucket=duplicate)
- Added a one-line note under the dropzone so the reviewer SEES the
  guarantee without needing to read the diff bucket pills first

CEO TODO: paste migrations/20260601_coin_event_bigint.sql into Supabase
SQL Editor (production). The Vercel deploy of this commit will still
fail to commit coin events until that migration runs — Prisma will
reject Int values that overflow.

+++ lib/chairops/pos-ingest/event-diff.ts
@@ -268,8 +268,12 @@ export async function ingestEvents(
         chairNumber: r.chairNumber,
         storeName: r.storeName,
         eventAt: r.eventAt,
-        coinAdded: Math.round(r.amount),
-        coinMeter: Math.round(r.meter),
+        // 2026-06-01: coinAdded/coinMeter are BIGINT in schema (StarThing
+        // exports unsigned 32-bit counters · int4 overflowed at ~4.29B).
+        // Math.round still works because incoming JS Number is well under
+        // 2^53; just wrap to BigInt for Prisma's createMany input.
+        coinAdded: BigInt(Math.round(r.amount)),
+        coinMeter: BigInt(Math.round(r.meter)),
         rowHash: r.rowHash,
         sourceImportId: importId,
       }));
